Fix windows build.

+++ editor/src/editor/systems/GizmoSystem.cpp
@@ -49,7 +49,7 @@ namespace se::editor::systems
         math::Vec3 mouseWorldPos = ScreenToWorldPoint(mousePos
            , cameraComp->view
            , cameraComp->proj
-           , math::Vec4(0, 0, viewportRect.size.x, viewportRect.size.y));
+           , math::Vec4(0.f, 0.f, static_cast<float>(viewportRect.size.x), static_cast<float>(viewportRect.size.y)));
         math::Vec3 direction = math::Normalized(mouseWorldPos - cameraComp->pos);
         geo::Ray ray = geo::Ray(cameraComp->pos, direction);
